lenme_task/loans/serializers.py
from rest_framework import serializers

from .models import Loan

class LoanSerializer(serializers.ModelSerializer):
    class Meta:
        model = Loan
        fields = '__all__'

class LoanCreateSerializer(serializers.ModelSerializer):
    class Meta:
        model = Loan
        fields = ('borrower', 'loan_amount', 'loan_period')
